new update
8

+++ my-app/src/components/PayStudentList.js
@@ -11,8 +11,8 @@ const PayStudentList = () => {
         email: '',
         contact: ''
     });
-    
-    
+
+
 
 
 
@@ -33,8 +33,8 @@ const PayStudentList = () => {
         fetchStudents();
     }, []);
 
-    
-    
+
+
     const handleSelectStudent = (studentId) => {
         setSelectedStudents(prevSelected => {
             const newSelected = new Set(prevSelected);
@@ -62,31 +62,31 @@ const PayStudentList = () => {
     };
     const renderImage = (imageText) => {
         if (imageText) {
-          const imageBytes = atob(imageText);
-          const arrayBuffer = new ArrayBuffer(imageBytes.length);
-          const uint8Array = new Uint8Array(arrayBuffer);
-          for (let i = 0; i < imageBytes.length; i++) {
-            uint8Array[i] = imageBytes.charCodeAt(i);
-          }
-          const blob = new Blob([uint8Array], { type: 'image/jpeg' });
-          const imageUrl = URL.createObjectURL(blob);
-          return <img src={imageUrl} alt="Student" style={{ maxWidth: '100px', maxHeight: '100px' }} />;
+            const imageBytes = atob(imageText);
+            const arrayBuffer = new ArrayBuffer(imageBytes.length);
+            const uint8Array = new Uint8Array(arrayBuffer);
+            for (let i = 0; i < imageBytes.length; i++) {
+                uint8Array[i] = imageBytes.charCodeAt(i);
+            }
+            const blob = new Blob([uint8Array], { type: 'image/jpeg' });
+            const imageUrl = URL.createObjectURL(blob);
+            return <img src={imageUrl} alt="Student" style={{ maxWidth: '100px', maxHeight: '100px' }} />;
         }
         return null;
-      };
-      const handlePaymentInitiation = () => {
+    };
+    const handlePaymentInitiation = () => {
         setShowModal(true);
     };
-    
+
     const handlePayment = async (userInfo) => {
         setShowModal(false); // Close the modal
         const amount = selectedStudents.size * 5; // Calculate total amount
         let order; // To store order details
-    
+
         try {
             const orderResponse = await axios.post('http://3.110.77.175:3000/createOrder', { amount });
             order = orderResponse.data; // Store order data
-    
+
             if (!window.Razorpay) {
                 const res = await loadScript('https://checkout.razorpay.com/v1/checkout.js');
                 if (!res) {
@@ -94,12 +94,12 @@ const PayStudentList = () => {
                     return;
                 }
             }
-    
+
             const options = {
-                key: 'rzp_test_e40XjG5y6SFo4h',
+                key: 'rzp_live_d4DgqU3P4V8cqL',
                 amount: order.amount.toString(),
                 currency: order.currency,
-                name: 'Vijayoc',
+                name: 'MSCE shorthand pune',
                 description: 'Test Transaction',
                 image: '',
                 order_id: order.id,
@@ -111,7 +111,7 @@ const PayStudentList = () => {
                         razorpay_signature: response.razorpay_signature,
                         studentIds
                     });
-                    
+
                     if (verificationResponse.data.success) {
                         fetchStudents();
                         alert('Payment successful!');
@@ -131,7 +131,7 @@ const PayStudentList = () => {
                     color: '#3399cc'
                 }
             };
-    
+
             const paymentWindow = new window.Razorpay(options);
             paymentWindow.open();
         } catch (error) {
@@ -139,18 +139,25 @@ const PayStudentList = () => {
             alert('Failed to initiate payment. Please try again.');
         }
     };
-    
-    
+
+
 
     const handleDeleteStudent = async (studentId) => {
-        try {
-            await axios.delete(`http://3.110.77.175:3000/studentsdel/${studentId}`);
-            fetchStudents();
-            // Refresh the student list after deletion
-        } catch (error) {
-            console.error('Failed to delete student:', error);
+        // Confirm before deletion
+        if (window.confirm("Are you sure you want to delete this student?")) {
+            try {
+                await axios.delete(`http://3.110.77.175:3000/studentsdel/${studentId}`);
+                fetchStudents(); // Refresh the student list after deletion
+                alert('Student deleted successfully');
+            } catch (error) {
+                console.error('Failed to delete student:', error);
+                alert('Failed to delete student. Please try again.');
+            }
+        } else {
+            alert('Deletion cancelled');
         }
     };
+    
 
     const buttonStyle = {
         padding: '12px 24px', // Generous padding for a larger click area
@@ -164,17 +171,17 @@ const PayStudentList = () => {
         boxShadow: '0px 2px 4px rgba(0,0,0,0.2)', // Subtle shadow for depth
         transition: 'background-color 0.2s, box-shadow 0.2s', // Smooth transition for hover effects
         margin: '20px'
-      };
-      
-      const buttonDisabledStyle = {
+    };
+
+    const buttonDisabledStyle = {
         ...buttonStyle,
         backgroundColor: '#9E9E9E', // A grey color indicating the button is disabled
         cursor: 'default', // Default cursor to indicate the button is not clickable
         boxShadow: 'none', // No shadow for a "flatter" disabled look
         margin: '20px'
-      };
+    };
 
-      const headingStyle = {
+    const headingStyle = {
         textAlign: 'center', // Centers the text
         color: '#333', // A modern, neutral color
         fontSize: '2rem', // A larger font size for the main heading
@@ -183,9 +190,9 @@ const PayStudentList = () => {
         letterSpacing: '1px', // Spacing out letters a bit for readability
         marginBottom: '1rem', // Adding some space below the heading
         paddingTop: '20px', // Padding at the top to push the content down a bit
-      };
+    };
 
-      const deleteButtonStyle = {
+    const deleteButtonStyle = {
         padding: '8px 16px',
         fontSize: '14px',
         fontWeight: 'bold',
@@ -197,91 +204,92 @@ const PayStudentList = () => {
         boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
         outline: 'none',
         transition: 'opacity 0.3s ease',
-      };
-      
-      const deleteButtonHoverStyle = {
+    };
+    const warningStyle  = {
+        color:'red'
+    }
+
+
+    const deleteButtonHoverStyle = {
         opacity: '0.8',
-      };
+    };
 
     return (
         <div className="student-list-container">
             <h1 style={headingStyle}>Students List</h1>
-            <p>After paying fees you cannot edit or delete student/शुल्क देण्यानंतर आपण विद्यार्थीचे संपादन किंवा हटवू शकत नाही.</p>
+            <p className='warning' style={warningStyle}>Delete facility is available only before payment. No edit / delete allowed once fees is Paid.
+डिलिट सुविधा फक्त फी भरण्यापूर्वी उपलब्ध आहे. एकदा फी भरल्यानंतर एडिट / डिलिट करता येणार नाही</p>
             <table className="student-table">
                 <thead>
                     <tr>
-                
+                        <th></th> {/* Empty header for the checkbox column */}
                         <th>Delete</th>
                         <th>Image</th>
                         <th>ID</th>
-                        <th>Password</th>
                         <th>Institute ID</th>
                         <th>Batch Start Date</th>
                         <th>Batch End Date</th>
-                        <th>First Name</th>
                         <th>Last Name</th>
-                        <th>Mother's Name</th>
+                        <th>First Name</th>
                         <th>Middle Name</th>
+                        <th>Mother's Name</th>
                         <th>Amount</th>
                         <th>Batch Year</th>
                         <th>Subject</th>
-        
                     </tr>
                 </thead>
                 <tbody>
                     {students.map(student => (
                         <tr key={student.student_id}>
                             <td>
-                            <button
-                                onClick={() => handleDeleteStudent(student.student_id)}
-                                style={deleteButtonStyle}
-                                onMouseOver={(e) => e.currentTarget.style.opacity = deleteButtonHoverStyle.opacity}
-                                onMouseOut={(e) => e.currentTarget.style.opacity = '1'}
-                                >
-                                Delete
-                            </button>
-                            </td>
-                            <td>{renderImage(student.image)}</td>
-                            <td>
                                 <input
                                     type="checkbox"
                                     checked={selectedStudents.has(student.student_id)}
                                     onChange={() => handleSelectStudent(student.student_id)}
-                                />{' '}
-                                {student.student_id}
+                                />
                             </td>
-                            <td>{student.password}</td>
+                            <td>
+                                <button
+                                    onClick={() => handleDeleteStudent(student.student_id)}
+                                    style={deleteButtonStyle}
+                                    onMouseOver={(e) => e.currentTarget.style.opacity = deleteButtonHoverStyle.opacity}
+                                    onMouseOut={(e) => e.currentTarget.style.opacity = '1'}
+                                >
+                                    Delete
+                                </button>
+                            </td>
+                            <td>{renderImage(student.image)}</td>
+                            <td>{student.student_id}</td>
                             <td>{student.instituteId}</td>
                             <td>{student.batchStartDate}</td>
                             <td>{student.batchEndDate}</td>
-                            <td>{student.firstName}</td>
                             <td>{student.lastName}</td>
-                            <td>{student.motherName}</td>
+                            <td>{student.firstName}</td>
                             <td>{student.middleName}</td>
+                            <td>{student.motherName}</td>
                             <td>{student.amount}</td>
                             <td>{student.batch_year}</td>
                             <td>{student.subject_name}</td>
-                            
                         </tr>
                     ))}
                 </tbody>
             </table>
             <div>{Array.from(selectedStudents).join(', ')}</div>
             <div className="payment-info">
-            <button
-                onClick={handlePaymentInitiation}
-                disabled={selectedStudents.size === 0}
-                style={selectedStudents.size === 0 ? buttonDisabledStyle : buttonStyle}
+                <button
+                    onClick={handlePaymentInitiation}
+                    disabled={selectedStudents.size === 0}
+                    style={selectedStudents.size === 0 ? buttonDisabledStyle : buttonStyle}
                 >
-                Pay ₹{selectedStudents.size * 5}
-            </button>
-        <PaymentModal
-            isOpen={showModal}
-            onClose={() => setShowModal(false)}
-            onSubmit={handlePayment}
-            userInfo={userInfo}
-            setUserInfo={setUserInfo}
-        />
+                    Pay ₹{selectedStudents.size * 5}
+                </button>
+                <PaymentModal
+                    isOpen={showModal}
+                    onClose={() => setShowModal(false)}
+                    onSubmit={handlePayment}
+                    userInfo={userInfo}
+                    setUserInfo={setUserInfo}
+                />
 
             </div>
         </div>
